% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.R
\name{add_data.Cell}
\alias{add_data.Cell}
\title{Method for adding data content to cells.}
\usage{
\method{add_data}{Cell}(item, content, field_name)
}
\arguments{
\item{item}{Item to add the formula to.}

\item{content}{String for the formula}

\item{field_name}{Name of the field to add the formula to (may be ignored).}

\item{location}{Used for recurssion, but might be ignored.}
}
\description{
Add content to a cell.
}
