% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.R
\name{add_data}
\alias{add_data}
\title{Method for adding data content.}
\usage{
add_data(item, content, field_name)
}
\arguments{
\item{item}{Item to add the formula to.}

\item{content}{String for the formula}

\item{field_name}{Name of the field to add the formula to (may be ignored).}
}
\description{
Adding content method.
}
