% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MondrianSheets.R
\name{add_fork_node}
\alias{add_fork_node}
\title{Method for adding a node to an excelable's layout connecting other nodes.}
\usage{
add_fork_node(x, sheet, id1, id2, axis)
}
\arguments{
\item{x}{\code{Excelable} object}

\item{sheet}{name of the sheet.}

\item{id1}{ID of the node 1}

\item{id2}{ID of the node 2}

\item{axis}{direction (row or column).}
}
\description{
Method for adding a node to an excelable's layout connecting other nodes.
}
