% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.R
\name{add_formula.Cell}
\alias{add_formula.Cell}
\title{Method for adding formulas to cells.}
\usage{
\method{add_formula}{Cell}(item, content, field_name, location = NULL)
}
\arguments{
\item{item}{Item to add the formula to.}

\item{content}{String for the formula}

\item{field_name}{Name of the field to add the formula to (may be ignored).}

\item{location}{Used for recurssion, but might be ignored.}
}
\description{
Formulas need to be added after the item is created, since the formula syntax allows for ommissions which assume reference to the object itself.
}
