% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MondrianSheets.R
\name{add_sheet.MondrianSheets}
\alias{add_sheet.MondrianSheets}
\title{Method for Excelable objects.}
\usage{
\method{add_sheet}{MondrianSheets}(x, name, label)
}
\arguments{
\item{x}{object}

\item{name}{name of the sheet.}

\item{label}{Label of the sheet.}
}
\description{
Method for Excelable objects.
}
