% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MondrianSheets.R
\name{add_sheet}
\alias{add_sheet}
\title{Method for adding a sheet to a layout.}
\usage{
add_sheet(x, name, label)
}
\arguments{
\item{x}{object}

\item{name}{name of the sheet.}

\item{label}{Label of the sheet.}
}
\description{
Method for adding a sheet to a layout.
}
