% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MondrianSheets.R
\name{add_terminal_node}
\alias{add_terminal_node}
\title{Method for adding an item to an excelable's layout as a terminal node.}
\usage{
add_terminal_node(x, sheet, item)
}
\arguments{
\item{x}{\code{Excelable} object}

\item{sheet}{name of the sheet.}

\item{item}{name of the item}
}
\description{
Method for adding an item to an excelable's layout as a terminal node.
}
