% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{apply_reference_schema_to_char}
\alias{apply_reference_schema_to_char}
\title{Internal function for converting a character reference into a proper
reference object using a ReferenceSchema}
\usage{
apply_reference_schema_to_char(x, reference_schema, location = NULL)
}
\arguments{
\item{x}{A character string.}

\item{reference_schema}{A ReferenceSchema}

\item{location}{Coordinate object indicating where the reference is.}
}
\description{
Internal function for converting a character reference into a proper
reference object using a ReferenceSchema
}
