% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.R
\name{Cell}
\alias{Cell}
\title{A Cell item object.}
\usage{
Cell(name, formatting)
}
\arguments{
\item{name}{Name of the item.}

\item{formatting}{list of formatting elements.}
}
\description{
A Cell item object.
}
