% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{check_char_against_reference_schema}
\alias{check_char_against_reference_schema}
\title{Internal function for converting a character reference into a proper
reference object using a ReferenceSchema}
\usage{
check_char_against_reference_schema(x, reference_schema)
}
\arguments{
\item{x}{A character string.}

\item{reference_schema}{A ReferenceSchema}
}
\description{
Internal function for converting a character reference into a proper
reference object using a ReferenceSchema
}
