% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MondrianSheets.R
\name{create_MondrianSheets}
\alias{create_MondrianSheets}
\title{Creator function for \verb{Mondrian Sheets}}
\usage{
create_MondrianSheets()
}
\description{
Creator function for \verb{Mondrian Sheets}
}
