% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excelable.R
\name{Excelable}
\alias{Excelable}
\title{Excelable object for creating the excel file with.}
\usage{
Excelable(item_set, layout)
}
\arguments{
\item{item_set}{An \code{ItemSet} object.}

\item{layout}{A layout object with defined methods.}
}
\description{
Excelable object for creating the excel file with.
}
