% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.R
\name{Formatting}
\alias{Formatting}
\title{Constructor for the formatting element.}
\usage{
Formatting(actions, formula = NULL)
}
\arguments{
\item{actions}{List of formatting actions.}

\item{formula}{Formula for conditional formatting.}
}
\description{
Not currently implemented.
}
