% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formula.R
\name{Formula}
\alias{Formula}
\title{Create a formula from a character string.}
\usage{
Formula(x, location)
}
\arguments{
\item{x}{Character string.}

\item{location}{Location the reference is being made from.}
}
\description{
Create a formula from a character string.
}
