% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.R
\name{new_Cell}
\alias{new_Cell}
\title{Constructor for the most basic item, a cell.}
\usage{
new_Cell(name, fields)
}
\arguments{
\item{name}{Name for the item.}

\item{fields}{A named list of the fields.}
}
\description{
Constructor for the most basic item, a cell.
}
