% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{new_Coordinate}
\alias{new_Coordinate}
\title{Constructor for a basic location.}
\usage{
new_Coordinate(item, field, column_start, column_end, row_start, row_end)
}
\arguments{
\item{item}{The name of the item the location is in.}

\item{field}{The name of the field the location is in.}

\item{column_start}{the column index (can be either character or numeric).}

\item{column_end}{the column index (can be either character or numeric).}

\item{row_start}{the row index (can be either character or numeric).}

\item{row_end}{the row index (can be either character or numeric).}
}
\description{
Constructor for a basic location.
}
