% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excelable.R
\name{new_Excelable}
\alias{new_Excelable}
\title{Constructor for Excelable object.}
\usage{
new_Excelable(item_set, layout)
}
\arguments{
\item{item_set}{An \code{ItemSet} object.}

\item{layout}{A layout object with defined methods.}
}
\description{
Constructor for Excelable object.
}
