% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.R
\name{new_Field}
\alias{new_Field}
\title{Constructor for a Field.}
\usage{
new_Field(content, formatting, row_names, column_names)
}
\arguments{
\item{content}{Takes the actual content, which is either a formula or data.}

\item{formatting}{Takes a list of formatting elements to be used for formatting.}

\item{row_names}{Names of the rows for references.}

\item{column_names}{Names of the columns for references.}
}
\description{
The point of a field is that items have a name and fields. There might need to be other types of field as well.
}
