% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell.R
\name{new_Formatting}
\alias{new_Formatting}
\title{Constructor for the formatting element.}
\usage{
new_Formatting(formula, actions)
}
\arguments{
\item{formula}{Formula for conditional formatting.}

\item{actions}{List of formatting actions.}
}
\description{
Constructor for the formatting element.
}
