% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formula.R
\name{new_Formula}
\alias{new_Formula}
\title{Constructor for the Formula class}
\usage{
new_Formula(statement, references)
}
\arguments{
\item{statement}{A string indicating the eventual excel formula.}

\item{references}{A list of Reference objects which correspond to the statement's placeholders.}
}
\description{
Constructor for the Formula class
}
