% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ItemSet.R
\name{new_ItemSet}
\alias{new_ItemSet}
\title{Constructor for the object which holds items.}
\usage{
new_ItemSet(items)
}
\arguments{
\item{items}{A list of items.}
}
\description{
Constructor for the object which holds items.
}
