% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MondrianSheets.R
\name{new_MondrianSheets}
\alias{new_MondrianSheets}
\title{Constructor for layout component \code{MondrianSheets}}
\usage{
new_MondrianSheets(sheets, layout)
}
\arguments{
\item{sheets}{data.frame of the sheets.}

\item{layout}{data.frame of the layout of items and nodes.}
}
\description{
Constructor for layout component \code{MondrianSheets}
}
