% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{new_Reference}
\alias{new_Reference}
\title{Constructor for referencing other items in a formula.}
\usage{
new_Reference(
  item,
  field,
  column_start,
  column_end,
  row_start,
  row_end,
  column_start_fixed,
  column_end_fixed,
  row_start_fixed,
  row_end_fixed
)
}
\arguments{
\item{item}{The name of the item being referenced.}

\item{field}{The name of the field within the item.}

\item{column_start}{The initial column index (can be either character or numeric).}

\item{column_end}{The final column index (can be either character or numeric).}

\item{row_start}{The initial row index (can be either character or numeric).}

\item{row_end}{The final row index (can be either character or numeric).}

\item{column_start_fixed}{Is the initial column index fixed?}

\item{column_end_fixed}{Is the final column index fixed?}

\item{row_start_fixed}{Is the initial row index fixed?}

\item{row_end_fixed}{Is the final row index fixed?}
}
\description{
Constructor for referencing other items in a formula.
}
