% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{new_ReferenceSchema}
\alias{new_ReferenceSchema}
\title{Constructor for defining types of references.}
\usage{
new_ReferenceSchema(pattern, extractions, implicits, equivalences)
}
\arguments{
\item{pattern}{Regular expression with extraction points.}

\item{extractions}{character vector indicating the reference element being extracted.}

\item{implicits}{named character vector indicating the implicits reference element values.}

\item{equivalences}{named character vector indicating where one element equals another, previously calculated element.}
}
\description{
Constructor for defining types of references.
}
