% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{Reference.character}
\alias{Reference.character}
\title{Cast a character to a Reference object}
\usage{
\method{Reference}{character}(x, location = NULL, ...)
}
\arguments{
\item{x}{An object}

\item{location}{Location the reference is being made from.}

\item{...}{Other parameter.}
}
\description{
Cast a character to a Reference object
}
