% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{Reference.list}
\alias{Reference.list}
\title{Cast a list to Reference}
\usage{
\method{Reference}{list}(x, location = NULL, ...)
}
\arguments{
\item{x}{An object}

\item{location}{Location the reference is being made from.}

\item{...}{Other parameter.}
}
\description{
Cast a list to Reference
}
