% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{Reference}
\alias{Reference}
\title{Method to cast objects to a Reference.}
\usage{
Reference(x, location = NULL, ...)
}
\arguments{
\item{x}{An object}

\item{location}{Location the reference is being made from.}

\item{...}{Other parameter.}
}
\description{
Method to cast objects to a Reference.
}
